fixed makefile: added "\RenderSystems" to include directories so that it compiles under windows. fixed main.cpp warning because of implicit castings. fixed system.cpp warning because of wrong order of initialization in constructor. updated driveline model (by benfsae): -changed Drive class to DriveMass -removed "commented" implementation of ODE masses for drive masses (in diff) -added 'gear' joint to simulate fixed joints and gears with a stiff torsional spring with dampening -recoded masses to use 'improved Euler' method for ODE solving -recoded joints to use velocity form for spring position and uses trapezoidal integration -removed Wheel::addTorque() {uses class definition} -removed DriveMass pointer functions (joints point to masses, not other way around) -renamed drive.hpp to driveMass.hpp -moved clutch.hpp data to driveJoint.hpp, fixed links -added LSD type joint: 1 input, 2 outputs (actually a spool right now) -change diff instance name to finalDrive in vehicle -removed transferW1, transferW2 and replaced with LSD diff

+++ src/main.cpp
@@ -158,7 +158,7 @@ void recordVideoFrames ()
 {
     static unsigned int time = 0;
     int takeShot = 0;
-    if ( time >= SystemData::getSystemDataPointer()->videoRecordTimestep )
+    if ( (int)time >= SystemData::getSystemDataPointer()->videoRecordTimestep )
     {
         takeShot++;
         time -= SystemData::getSystemDataPointer()->physicsTimeStep;
